added new security in album and statistic

+++ src/Application/Services/AlbumService.cs
@@ -1,7 +1,6 @@
 using Application.DTOs.Album.Request;
 using Application.DTOs.Album.Response;
 using Application.Interfaces;
-using Application.Validators;
 
 using Domain.Entities;
 using Domain.Interfaces;
@@ -12,11 +11,11 @@ namespace Application.Services;
 public class AlbumService : IAlbumService
 {
     private IAlbumRepository _album;
-    private IUserRepository _user;
+    private IUserService _user;
     private readonly ISongRepository _song;
 
 
-    public AlbumService(IAlbumRepository album, ISongRepository song, IUserRepository user)
+    public AlbumService(IAlbumRepository album, ISongRepository song, IUserService user)
     {
         _album = album;
         _song = song;
@@ -183,12 +182,11 @@ public class AlbumService : IAlbumService
             throw new NotFoundException("Album");
         }
 
-        // if (idUser !=  album.IdArtist || user.Role != "Admin")
-        // {
-        //     throw new Exception("You don't have permission to delete this album");
-        // }
+        if (idUser !=  album.IdArtist || user.Role != "Admin")
+        {
+            throw new Exception("You don't have permission to delete this album");
+        }
 
-        AuthValidator.ValidateAdminOrOwner(album.IdArtist, idUser, user.Role!, "You don't have permission to delete this album");
 
 
 
@@ -220,22 +218,35 @@ public class AlbumService : IAlbumService
             throw new NotFoundException("Song");
         }
 
-        // if(user.Id != idUser)
-        // {
-        //     throw new Exception("You don't have permission to add this song");
-        // }
-        AuthValidator.ValidateOwner(album.IdArtist, idUser, "You don't have permission to add this song");
-        AuthValidator.ValidateOwner(song.IdArtist, idUser, "You don't have permission to add this song");
+
+
+        if(user.Id != idUser)
+        {
+            throw new Exception("You don't have permission to add this song");
+        }
+        // AuthValidator.ValidateOwner(album.IdArtist, idUser, "You don't have permission to add this song");
+        // AuthValidator.ValidateOwner(song.IdArtist, idUser, "You don't have permission to add this song");
 
         //
 
-        // if(user.Id != song.IdArtist)
-        // {
-        //     throw new Exception("You don't have permission to add this song");
-        // }
+        if(user.Id != song.IdArtist)
+        {
+            throw new Exception("You don't have permission to add this song");
+        }
+
+        if(song.IdAlbum != Guid.Empty)
+        {
+            throw new Exception("The song is already in an album");
+        }
+
+        if(song.IdAlbum == album.Id)
+        {
+            throw new Exception("The song is already in this album");
+        }
 
 
         album.AddSong(song);
+        
         await _album.Update(album, cancellationToken);
 
         return new GetByIdResponse
@@ -269,15 +280,16 @@ public class AlbumService : IAlbumService
             throw new NotFoundException("Song");
         }
 
-        // if(user.Id != idUser || user.Role != "Admin" )
-        // {
-        //     throw new Exception("You don't have permission to delete this song");
-        // }
-        AuthValidator.ValidateAdminOrOwner(album.IdArtist, idUser, user.Role!, "You don't have permission to delete this song");
+        if(user.Id != idUser && user.Role != "Admin" )
+        {
+            throw new Exception("You don't have permission to delete this song");
+        }
+        // AuthValidator.ValidateAdminOrOwner(album.IdArtist, idUser, user.Role!, "You don't have permission to delete this song");
         
 
 
         album.DeleteSong(song);
+        song.RemoveFromAlbum();
         await _album.Update(album, cancellationToken);
 
         return new GetByIdResponse(
